Update bulk notification request DTO

+++ notification-service/src/main/java/com/quickbite/notification/dto/BulkNotificationRequest.java
@@ -12,7 +12,7 @@ import lombok.Data;
 public class BulkNotificationRequest {
 
     @NotEmpty(message = "recipientIds must not be empty")
-    private List<@NotNull @Min(value = 1, message = "recipientId must be greater than 0") Integer> recipientIds;
+    private List<@NotNull @Min(value = 1, message = "recipientId must be greater than 0") Long> recipientIds;
 
     @Valid
     @NotNull(message = "notification payload is required")
